wr-136: * move user info block to separate component
