+++ frontend/src/pages/profile/styles.module.scss
@@ -1,36 +1,16 @@
 .profile {
+  display: grid;
+  grid-template: "info activity activity" 1.4fr "info articles genres" 0.3fr "achievements articles genres" 1.3fr / minmax(
+      270px,
+      0.5fr
+    ) 1.3fr 1.2fr;
+  gap: 30px;
   width: 100%;
   max-width: 1590px;
   margin: 0 auto;
 }
 
-.userInfoBlock {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 20px;
-  width: 100%;
-  max-width: 270px;
-  padding: 20px 30px;
-  text-align: center;
+.profileBlock {
   border-radius: 7px;
-  box-shadow: 0 4px 8px -2px rgb(16 24 40 / 16%);
-}
-
-.avatar {
-  width: 80%;
-  aspect-ratio: 1 / 1;
-  height: auto;
-}
-
-.userInfo {
-  font-size: 18px;
-}
-
-.editProfileBtn {
-  width: 100%;
-  height: 45px;
-  color: var(--white);
-  background-color: var(--dark-green);
-  border: none;
+  box-shadow: 0 0 6px rgb(16 24 40 / 16%);
 }
